Update README.md
I added a little clause in the "HOW TO INSTALL UCLATHES" section explaining that the "directory where you run LaTeX" into which the .cls, .clo, and .bst files ought to go is the same folder where your dissertation's / thesis's .tex file live. 

Maybe it's just me, but this clarification seems crucial. As a relative LaTeX novice (not to mention a GitHub novice) and someone not used to installing packages manually (i.e. outside of a package manager like MikTeX), figuring this out late one night was a huge game-changer, as previously I'd been trying to put these files in the root latex directories on my machine and having zero luck when trying to compile \documentclass{uclathes}. 

Once I cracked the code on this, uclathes runs beautifully. Go Bruins.